requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
easydict==1.13
filelock==3.16.1
flatbuffers==25.1.24
fonttools==4.56.0
fsspec==2024.12.0
gast==0.4.0
gdown==5.2.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.5
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.41.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
MNN==3.0.3
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
onnx==1.17.0
onnx-tf==1.10.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
openvino==2023.3.0
openvino-telemetry==2025.0.0
opt_einsum==3.4.0
packaging==24.2
pillow==10.4.0
proglog==0.1.10
protobuf==3.20.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
six==1.17.0
soupsieve==2.6
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-addons==0.21.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.4.0
tf2onnx==1.16.1
thop==0.1.1.post2209072238
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
